requirements.txt
Flask==2.3.2
flask-restful==0.3.10
python-dotenv==1.0.0
mysql-connector-python==8.1.0